requirements.txt
moviepy==1.0.3
numpy>=1.19.5
tk>=0.1.0
ffmpeg-python>=0.2.0
pillow<=9.0.0 